requirements.txt
pyyaml>=6.0
pandas>=1.2.4
numpy>=1.22.3
flask>=2.0.3
xlrd==2.0.1
psycopg2==2.8.6
openpyxl==3.0.9
flask-login>=0.5.0